validation for deleting product only seller
- a product could be deleted only seller

+++ app/controllers/products_controller.rb
@@ -1,4 +1,6 @@
 class ProductsController < ApplicationController
+  before_action :product_seller?, only: :item
+
   def index
     @products = Product.order('created_at DESC').includes(:images)
   end
@@ -84,19 +86,21 @@ class ProductsController < ApplicationController
     
   def destroy
     @product = Product.find(params[:id])
-    if @product.destroy
-      redirect_to my_selling_products_users_path, notice: "商品を削除しました"
+    if @product.seller == current_user
+      if @product.destroy
+      redirect_to my_selling_products_users_path, notice: "商品を削除しました" and return
+      else
+        render :item, collection: @product and return
+      end
     else
-      render :item, collection: @product
+      redirect_to root_path and return
     end
   end
 
   def item
     @category = []
     @product = Product.find(params[:id])
-    @price = (@product.price * 1.08).ceil
     @category = @product.categories.pluck(:name)
-    @seller = @product.seller
   end
     
   private
@@ -116,6 +120,11 @@ class ProductsController < ApplicationController
       category_ids: []
     )
     .merge(seller_id: current_user.id)
+  end
+
+  def product_seller?
+    @product = Product.find(params[:id])
+    redirect_to root_path unless @product.seller == current_user 
   end
   
 end
